rename pkg lsp to langserver
