+++ langserver/handle_text_document.go
@@ -1,4 +1,4 @@
-package lsp
+package langserver
 
 import (
 	"bytes"
