+++ langserver/handle_text_document_test.go
@@ -1,4 +1,4 @@
-package lsp
+package langserver
 
 import (
 	"reflect"
