+++ langserver/main_test.go
@@ -1,4 +1,4 @@
-package lsp
+package langserver
 
 import (
 	"flag"
