+++ main.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"net"
 
-	"github.com/blockchain-labs-org/solzaemon/lsp"
+	"github.com/blockchain-labs-org/solzaemon/langserver"
 	"github.com/sourcegraph/jsonrpc2"
 )
 
@@ -13,13 +13,13 @@ var connOpt = []jsonrpc2.ConnOpt{}
 
 func main() {
 	log.Printf("listen :8080")
-	err := launch(lsp.NewHandler(), ":8080")
+	err := launch(langserver.NewHandler(), ":8080")
 	if err != nil {
 		panic(err)
 	}
 }
 
-func launch(handler *lsp.Handler, addr string) error {
+func launch(handler *langserver.Handler, addr string) error {
 	h := jsonrpc2.HandlerWithError(func(ctx context.Context, conn *jsonrpc2.Conn, req *jsonrpc2.Request) (result interface{}, err error) {
 		return handler.Handle(ctx, conn, req)
 	})
